+++ app/shared/data.js
@@ -15,6 +15,7 @@ define(() => {
     let easyrtcIds = [];
     let language = '';
     let langJson = {};
+    let roomDraft = {};
 
     const useAdapterForRoom = (room) => ({
         channel: room.channel,
@@ -115,6 +116,11 @@ define(() => {
         getLangJson: () => langJson,
         setLangJson: (value) => {
             langJson = value;
+        },
+
+        getRoomDraft: () => roomDraft,
+        setRoomDraft: (value) => {
+            roomDraft = value;
         }
     };
 });