+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All changes to XM WEB CLIENT project will be documented in this file.
 ### Add
 - Add Multiple Language.
 - Add message immediately after send and display its status.
+- Add draft message.
 
 ### [3.6] - 2021-03-24
 ### Fix